src/main/java/com/example/dotcode/service/impl/UserServiceImpl.java
package com.example.dotcode.service.impl;

import com.example.dotcode.entity.dto.UserDto;
import com.example.dotcode.entity.mapper.UserMapper;
import com.example.dotcode.repository.UserRepository;
import com.example.dotcode.service.UserService;
import org.springframework.stereotype.Service;

import java.util.List;


@Service
public class UserServiceImpl implements UserService {

    private final UserRepository userRepository;
    private final UserMapper userMapper;

    public UserServiceImpl(UserRepository userRepository, UserMapper userMapper) {
        this.userRepository = userRepository;
        this.userMapper = userMapper;
    }

    @Override
    public List<UserDto> getAll() {
        return userRepository.findAll()
                .stream()
                .map(userMapper::toDto)
                .toList();
    }

    @Override
    public UserDto save(UserDto userDto) {
        return userMapper.toDto(userRepository.save(userMapper.toEntity(userDto)));
    }

    @Override
    public UserDto update(int id, UserDto userDto) {
        if (!userRepository.existsById(id)) {
            throw new IllegalArgumentException("User with ID " + userDto.getId() + " does not exist.");
        }
        return userMapper.toDto(userRepository.save(userMapper.toEntity(userDto)));
    }

    @Override
    public Integer delete(int id) {
        userRepository.deleteById(id);
        return id;
    }

}

